refactor: move readmodel daily stats packages
