+++ pearljam-api/src/main/java/fr/insee/pearljam/api/reporting/presenter/CampaignCollectionPresenter.java
@@ -5,7 +5,7 @@ import fr.insee.pearljam.domain.reporting.port.in.CampaignStatsPresenter;
 import fr.insee.pearljam.api.reporting.response.ClosingCausesProgressResponse;
 import fr.insee.pearljam.api.reporting.response.CollectionRatesResponse;
 import fr.insee.pearljam.api.reporting.response.ContactOutcomesProgressResponse;
-import fr.insee.pearljam.domain.reporting.readmodel.stats.CampaignDailyStats;
+import fr.insee.pearljam.domain.reporting.readmodel.CampaignDailyStats;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
