+++ pearljam-api/src/main/java/fr/insee/pearljam/api/reporting/presenter/CampaignProgressByInterviewersPresenter.java
@@ -4,8 +4,8 @@ import fr.insee.pearljam.api.reporting.response.CampaignProgressByInterviewersRe
 import fr.insee.pearljam.domain.reporting.port.in.CampaignStatsByInterviewersPresenter;
 import fr.insee.pearljam.api.reporting.response.CommunicationsProgressResponse;
 import fr.insee.pearljam.api.reporting.response.StatesProgressResponse;
-import fr.insee.pearljam.domain.reporting.readmodel.stats.CampaignDailyStats;
-import fr.insee.pearljam.domain.reporting.readmodel.stats.InterviewerDailyStats;
+import fr.insee.pearljam.domain.reporting.readmodel.CampaignDailyStats;
+import fr.insee.pearljam.domain.reporting.readmodel.InterviewerDailyStats;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
