+++ pearljam-api/src/main/java/fr/insee/pearljam/api/reporting/presenter/CampaignProgressPresenter.java
@@ -4,7 +4,7 @@ import fr.insee.pearljam.api.reporting.response.CampaignProgressResponse;
 import fr.insee.pearljam.domain.reporting.port.in.CampaignStatsPresenter;
 import fr.insee.pearljam.api.reporting.response.CommunicationsProgressResponse;
 import fr.insee.pearljam.api.reporting.response.StatesProgressResponse;
-import fr.insee.pearljam.domain.reporting.readmodel.stats.CampaignDailyStats;
+import fr.insee.pearljam.domain.reporting.readmodel.CampaignDailyStats;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
