+++ pearljam-api/src/main/java/fr/insee/pearljam/api/reporting/response/CampaignCollectionByInterviewersResponse.java
@@ -1,7 +1,7 @@
 package fr.insee.pearljam.api.reporting.response;
 
-import fr.insee.pearljam.domain.reporting.readmodel.stats.CampaignDailyStats;
-import fr.insee.pearljam.domain.reporting.readmodel.stats.InterviewerDailyStats;
+import fr.insee.pearljam.domain.reporting.readmodel.CampaignDailyStats;
+import fr.insee.pearljam.domain.reporting.readmodel.InterviewerDailyStats;
 
 import java.util.List;
