+++ pearljam-domain/src/test/java/fr/insee/pearljam/domain/campaign/stub/CampaignDailyStatsRepositoryPortStub.java
@@ -1,9 +1,9 @@
 package fr.insee.pearljam.domain.campaign.stub;
 
 import fr.insee.pearljam.domain.reporting.port.out.CampaignDailyStatsRepositoryPort;
-import fr.insee.pearljam.domain.reporting.readmodel.stats.CampaignDailyStats;
-import fr.insee.pearljam.domain.reporting.readmodel.stats.InterviewerDailyStats;
-import fr.insee.pearljam.domain.reporting.readmodel.stats.OrganizationUnitDailyStats;
+import fr.insee.pearljam.domain.reporting.readmodel.CampaignDailyStats;
+import fr.insee.pearljam.domain.reporting.readmodel.InterviewerDailyStats;
+import fr.insee.pearljam.domain.reporting.readmodel.OrganizationUnitDailyStats;
 
 import java.time.LocalDate;
 import java.util.List;
